refactor: Entity Converter 개선
