+++ app/src/main/java/woowacourse/movie/mapper/TicketEntityMapper.kt
@@ -2,7 +2,6 @@ package woowacourse.movie.mapper
 
 import woowacourse.movie.data.entity.TicketEntity
 import woowacourse.movie.model.HeadCount
-import woowacourse.movie.model.Seat
 import woowacourse.movie.model.Seats
 import woowacourse.movie.model.Ticket
 
@@ -12,7 +11,7 @@ fun Ticket.toEntity(): TicketEntity {
         date = selectedDate,
         time = selectedTime,
         headCount = headCount.value,
-        seat = convertSeat(seats.seats),
+        seat = seats.seats,
         theater = theater,
         price = amount,
     )
@@ -25,23 +24,6 @@ fun TicketEntity.toDomain(): Ticket {
         headCount = HeadCount(headCount),
         selectedDate = date,
         selectedTime = time,
-        seats = Seats(parseSeats(seat)),
+        seats = Seats(seat),
     )
-}
-
-private fun convertSeat(seats: Set<Seat>): String {
-    return seats.joinToString(", ") { point ->
-        "${'A' + point.row}${point.col + 1}"
-    }
-}
-
-private fun parseSeats(seatString: String): Set<Seat> {
-    return seatString.split(", ")
-        .map { seat ->
-            val rowChar = seat[0]
-            val colNumber = seat.substring(1)
-            val row = rowChar - 'A'
-            val col = colNumber.toInt() - 1
-            Seat(row, col)
-        }.toSet()
 }
